requirements.txt
# Core API dependencies
google-generativeai>=0.3.0
openai>=1.0.0
anthropic>=0.8.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
datasets>=2.14.0
huggingface-hub>=0.19.0
transformers

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.66.0
pyyaml>=6.0
jsonlines>=4.0.0

# Testing 
pytest>=7.4.0
nltk
